pyproject.toml

# pyproject.toml - Configuration projet Python

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ui-pro"
version = "0.1.0"
description = "LLM orchestration and memory management"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1>=0",
    "requests>=2>=1",
    "beautifulsoup4>=4>=0",
    "faiss-cpu>=1>=0",
    "numpy>=1>=0",
    "sentence-transformers>=2>=0",
    "python-multipart>=0>=0",
    "uvicorn>=0>=0",
    "fastapi>=0>=0",
    "jinja2>=3>=0",
    "pyyaml>=6>=0",
    "pytest>=7>=0",
    "pytest-mock>=3>=0",
]

[project.optional-dependencies]
dev = [
    "black>=23",
    "isort>=5",
    "flake8>=6",
    "mypy>=1",
]

[tool.black]
# Black configuration
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | venv
    | build
    | dist
)/'''

[tool.isort]
# isort configuration
profile = "black"
line_length = 88
skip = [
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
]
# Multi-line output
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
# Mypy configuration - type checking
strict = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = false
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["fastapi.*,pydantic.*,requests.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
# Pytest configuration
testpaths = ["tests"]
addopts = "-v --cov=ui-pro --cov-report=term --cov-report=xml"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
branch = true
source = ["ui-pro"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_patterns = [
    "*/tests/*",
    "*/__init__.py",
]
